+++ test/unit/items-test.js
@@ -22,13 +22,26 @@ exports.examples = (expect) => {
       });
     });
 
-    describe('.configureQueue(messageQueueParams)', () => {
+    describe('.configureQueue(filterParams)', () => {
       it('PUTs to /data/v1/items/queues with messageQueueParams as the body', function () {
-        const messageQueueParams = { fromZone: 'exitZone', toZone: 'entryZone' };
-        return expect(this.subject.items.configureQueue(messageQueueParams)).to.wrap.request({
+        const filterParams = { fromZone: 'exitZone', toZone: 'entryZone' };
+        return expect(this.subject.items.configureQueue(filterParams)).to.wrap.request({
           method: 'put',
           path: '/itemsense/data/v1/items/queues',
-          body: messageQueueParams
+          body: filterParams
+        });
+      });
+    });
+
+    describe('.configureQueue(filterParams, options)', () => {
+      it('PUTs to /data/v1/items/queues with messageQueueParams as the body', function () {
+        const filterParams = { fromZone: 'exitZone', toZone: 'entryZone' };
+        const optionParams = { deliveryMethod: 'NON_PERSISTENT' };
+        return expect(this.subject.items.configureQueue(filterParams, optionParams)).to.wrap.request({
+          method: 'put',
+          path: '/itemsense/data/v1/items/queues',
+          body: filterParams,
+          query: optionParams
         });
       });
     });